backend/requirements.txt
Flask
Flask-Cors
openai
langchain_community
langchain 
langchain_openai
pypdf
# dotenv
chromadb